requirements.txt
streamlit
google-generativeai
python-dotenv
langchain_community
PyPDF2
langchain_google_genai
faiss-cpu